refactor(inbox): extract the Needs-review column (#994)
Moves the unclassified-file table, select-all row, bulk-override fieldset
(FR-044/US11, #755), the #611 heterogeneous gate and undo banner, and the
per-file apply row into InboxNeedsReview.tsx, along with the derivations
only that surface used (genericBulkFields, KNOWN_BULK_FIELD_LABELS,
allSelected/someSelected).

The component takes the whole useInboxReclassifyState result rather than
20 separate props, so the boundary stays stable as the hook grows.

The empty case stays a call-site ternary: TwoColDetailLayout allocates a
`__col` per non-null extraCols entry, so a component that merely rendered
null internally still claimed a column and broke the #813 layout test.

InboxDetail.tsx: 930 -> 627 lines.
